Add step to check file contents when updating a dataset
When updating a dataset, if the hashes (SHA256 and MD5) and file type (Raw, Columnar/TableCSV, HDF5/NumericMatrixCSV) of a local file match a datafile in the base dataset version, add the existing datafile as a virtual datafile instead of uploading the local file.

Scripts that upload a batch of files may only be changing one at a time, especially in the case of fixing mistakes. It would be helpful if the TaigaClient didn't upload these duplications. The benefits would be:
- Faster upload
- Easier to tell which files actually changed in the Taiga UI when looking at the change log
- Less wasted space