feat: get_datafile_metadata now returns original_file_sha256
get_datafile_metadata() returns instances of MinDataFileMetadata which
now have a field named `original_file_sha256`. This hash is the sha256 hash
of the file that was originally uploaded to Taiga and is useful for checking to
see whether a file in Taiga is the same as one you have locally. (For example,
checking before uploading a file. If the file is the same, no need to upload.)